requirements.txt
flask
gunicorn
pymongo[srv]
python-dotenv
passlib
flask-wtf
flask-security
email_validator